create trajectories based on probabilities

+++ src/main/java/com/terator/model/generatorTable/ProbabilitiesAndNumberOfDrawsFromBuildingInSpecificTime.java
@@ -6,6 +6,6 @@ import java.util.Map;
 
 public record ProbabilitiesAndNumberOfDrawsFromBuildingInSpecificTime(
         Map<BuildingType, Double> probabilityToType,
-        Long numberOfDraws
+        int numberOfDraws
 ) {
 }
